Refactored location of IMultiSourceTargetProvider
