+++ src/main/java/com/hollingsworth/arsnouveau/api/source/IMultiSourceTargetProvider.java
@@ -1,4 +1,4 @@
-package com.hollingsworth.arsnouveau.common.block.tile;
+package com.hollingsworth.arsnouveau.api.source;
 
 import net.minecraft.core.BlockPos;
